% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iccbot.R
\name{compute_icc}
\alias{compute_icc}
\title{This is the core of the irr::icc() function}
\usage{
compute_icc(ns, nr, MSr, MSw, MSc, MSe, model, type, unit, r0, conf.level)
}
\description{
This is the core of the irr::icc() function
}
